configs: usbprod: Update all to use GNU coreutils
It has been observed that Busybox md5sum cannot handle certain
backslash-escaped filenames. This is a problem for our scripts
which use md5sums to do a file check after unpacking a tarball.
This problem popped up for now with Debian Bullseye based
distributions, it has not been observed with other distributions.