tsimx6ul-usbprod: Updated blast process
Adds an image capture process, if no images are present on
the USB drive once it is booted, the script will capture images
from any bootable media. The source is undisturbed, a copy is made,
the filesystem sanitized (based on our prep_customer_image scripts),
and either a tarball or disk image is created. The USB drive can then
be inserted directly in to another device of the same model, and those
images will be written to bootable media on the target.

This update also streamlines the blast process and cleans up scripting.
All of the write/verification process is handled by blast_funcs.sh,
which can be called as needed for easily customizing the blast process.
The blast.sh script is platform dependent and defines device nodes, etc.