usbprod: Pass '-T default' to mke2fs
It was found that, in creating smaller disk images, 64-128 MB,
mke2fs was trying to be clever and use a smaller blocksize. This
in turn caused extremely long load times (10s to 100s of seconds)
in U-Boot while loading the .cpio files from USB. Since these
disk images are also set up to expand, the sensible thing to do
is tell mke2fs to use a different profile for more sane options
to the filesystem.