package/ts7400v2-uboot: Update to support Image Replicator
U-Boot change that runs the usbprod script at startup to
facilitate running of the Image Replicator tool on this platform.

The TS-7400-V2 also needed a change to turn USB power on at usbprod
startup. This is normally handled by the bootloader but was not
implemented in the U-Boot port originally.