blast_funcs: modify tarball capture to not grab disk image
Currently, the image capture process images the whole source disk
and then performs sanitization operations on it before making a
tarball of the contents.

This changes that paradigm by creating a sparse file backed loopback
on the destination path, formats it, copies the filesystem contents,
and then performs the normal operations on it. This greatly reduces
USB requirements and run time for capturing an image from a source
disk with a single partition.