blast_funcs: Only decompress once when writing dd
Previously, the source file would be decompressed twice, once to
write it to disk and the second to run through wc to get byte
count. Using a FIFO, pipe the decompressed data to wc and dd when
writing to disk. Byte count can be grabbed later if doing md5 verify.