usbprod-common/blast_funcs.sh: Use pipefile on script
In testing there was a situation where the USB device failed but
ultimately showed a success because the stream decompression failed,
which was piped in to a write command that succeeded. The rest of
the script got gummed up because now the .md5 file disappeared so
the script just marked the whole process as successful while the
actual image written to disk was bad.

Enabling pipefile in the script will correctly fail in similar
situations.

Signed-off-by: Kris Bahnsen <kris@embeddedTS.com>