tsimx6-usbprod: blash.sh: Fix incorrect path variable
A copy/paste error led to a wrong variable name being used. This
is very unlikely to have shown up in the field as the only time
this part of the script would have been reached is if a user was
attempting to install a very old U-Boot release.

This also cleans up the eval command and unsets imx_type before
checking. There was an issue with imx_type being still set in the
script that could cause an incorrect U-Boot binary to be approved
and installed.

Fixes #39