usbprod: Touch /ssh_regenkeys in sanitize_linux_rootfs.sh
Some newer distro-boot startup scripts look for this as a trigger
rather than /firstboot. This will be fixed in distro-boot to
automatically regen keys if they don't already exist. This is to
handle any existing images that may be in the wild.

As a result, this may leave some empty files in the final filesystem.

Signed-off-by: Kris Bahnsen <kris@embeddedTS.com>