blast_funcs: add support for option flags
Fixes #27

Creates a function that can be used to search for files in a folder
that match the name "IR_*" and these are then exported as environment
variables.

While a few of them are now documented, blast_funcs.sh currently
will only expect IR_NO_COMPRESS and act on that.

Signed-off-by: Kris Bahnsen <kris@embeddedTS.com>