Image Replicator: make LED control less manual
In 890353b, nearly all Image Replicator configs were bumped to 6.6
from 4.9. The 6.6 kernel changed how LEDs should be named which then
broke all of the Image Replicator tools.

The LED setup and manipulation was fully moved to blast_funcs.sh
and individual setup removed from platform blash.sh scripts. The
new setup and manipulation automates looking for the LEDs using the
newer kernel naming paradigm.

This does lose the ability for platforms to use their own arbitrary
control of LEDs; e.g. using tshwctl, which has been deprecated in
favor of kernel control for a while. So this loss should not be a
problem.

Fixes: 890353b ("configs: enable additional SSL engines, bump to v6.6.84-ts")
Signed-off-by: Kris Bahnsen <kris@embeddedTS.com>